Move User to the database module.
