+++ database/user_test.go
@@ -1,4 +1,4 @@
-package user
+package database
 
 import "testing"
